Small projects updates
