+++ src/Components/Projects.jsx
@@ -1,0 +1,13 @@
+import "./Component-Styles/Projects.css";
+
+function Projects() {
+  return (
+    <div className="projects-container">
+      <section id="projects">
+      <h2 className="projects-title">My projects</h2>
+      </section>
+    </div>
+  );
+}
+
+export default Projects;
